Fix check for superfences config when loaded via pymdownx.extra
The check for the `superfences` config failed when `superfences` was loaded by `pymdownx.extra`. In particular, this means that the plugin failed to work with the defaults set up by Spotify's `techdocs-core` plugin. In this PR, the check is adapted to also check for the `pymdownx.superfences` configuration inside the `pymdownx.extra` configuration.